[Engine] Implement Sys_Error, DCV_FB_Glyph and cmd.c; fix DC 16-bit datatypes; build at /O2 /Ob2

+++ src/engine/common.h
@@ -4,7 +4,7 @@
 #define COMMON_H
 #pragma once
 
-typedef int qboolean;
+typedef short qboolean;
 
 //============================================================================
